Added metric tests
